Move related utilities to new dom and url package
