+++ packages/dom/README.md
@@ -1,0 +1,3 @@
+# DOM
+
+DOM utility package for Web Stories editor.
