+++ packages/dom/src/escapeHTML.js
