+++ packages/dom/src/getCaretCharacterOffsetWithin.js
