+++ packages/dom/src/index.js
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 Google LLC
+ * Copyright 2022 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,10 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
-const buffer = document.createElement('div');
-export default function stripHTML(string) {
-  // @todo: implement a cheaper way to strip markup.
-  buffer.innerHTML = string;
-  return buffer.textContent;
-}
+export { default as stripHTML } from './stripHTML';
+export { default as escapeHTML } from './escapeHTML';
+export { default as getCaretCharacterOffsetWithin } from './getCaretCharacterOffsetWithin';
+export { default as loadStylesheet } from './loadStylesheet';
