+++ packages/dom/src/test/escapeHTML.js
