+++ packages/story-editor/package.json
@@ -36,6 +36,7 @@
     "@googleforcreators/animation": "*",
     "@googleforcreators/date": "*",
     "@googleforcreators/design-system": "*",
+    "@googleforcreators/dom": "*",
     "@googleforcreators/fonts": "*",
     "@googleforcreators/i18n": "*",
     "@googleforcreators/media": "*",
@@ -50,6 +51,7 @@
     "@googleforcreators/tracking": "*",
     "@googleforcreators/transform": "*",
     "@googleforcreators/units": "*",
+    "@googleforcreators/url": "*",
     "blurhash": "^1.1.4",
     "classnames": "^2.3.1",
     "clone-deep": "^4.0.1",
