+++ packages/story-editor/src/components/elementLink/output.js
@@ -18,7 +18,7 @@
  * External dependencies
  */
 import PropTypes from 'prop-types';
-import { withProtocol } from '@googleforcreators/design-system';
+import { withProtocol } from '@googleforcreators/url';
 
 /**
  * Internal dependencies
