+++ packages/story-editor/src/components/library/panes/media/local/hotlink/hotlinkModal.js
@@ -17,7 +17,8 @@
  * External dependencies
  */
 import { __, sprintf, translateToExclusiveList } from '@googleforcreators/i18n';
-import { Input, withProtocol } from '@googleforcreators/design-system';
+import { withProtocol } from '@googleforcreators/url';
+import { Input } from '@googleforcreators/design-system';
 import {
   useState,
   useRef,
