+++ packages/story-editor/src/components/panels/design/textStyle/stylePresets/styleItem.js
@@ -22,12 +22,13 @@ import PropTypes from 'prop-types';
 import { __ } from '@googleforcreators/i18n';
 import { generatePatternStyles } from '@googleforcreators/patterns';
 import { Icons } from '@googleforcreators/design-system';
+import { stripHTML } from '@googleforcreators/dom';
+
 /**
  * Internal dependencies
  */
 import { BACKGROUND_TEXT_MODE } from '../../../../../constants';
 import { generatePresetStyle } from '../../../../../utils/presetUtils';
-import stripHTML from '../../../../../utils/stripHTML';
 import { useStory } from '../../../../../app/story';
 import { focusStyle } from '../../../shared';
