+++ packages/story-editor/src/elements/text/display.js
@@ -24,6 +24,7 @@ import { createSolid } from '@googleforcreators/patterns';
 import { useUnits } from '@googleforcreators/units';
 import { useTransformHandler } from '@googleforcreators/transform';
 import { getHTMLFormatters, getHTMLInfo } from '@googleforcreators/rich-text';
+import { stripHTML } from '@googleforcreators/dom';
 
 /**
  * Internal dependencies
@@ -39,7 +40,6 @@ import {
 } from '../shared';
 import StoryPropTypes from '../../types';
 import { BACKGROUND_TEXT_MODE } from '../../constants';
-import stripHTML from '../../utils/stripHTML';
 import {
   getResponsiveBorder,
   shouldDisplayBorder,
