+++ packages/story-editor/src/elements/text/edit.js
@@ -38,6 +38,7 @@ import {
   getHTMLInfo,
   useRichText,
 } from '@googleforcreators/rich-text';
+import { stripHTML } from '@googleforcreators/dom';
 
 /**
  * Internal dependencies
@@ -53,7 +54,6 @@ import {
 } from '../shared';
 import StoryPropTypes from '../../types';
 import { BACKGROUND_TEXT_MODE } from '../../constants';
-import stripHTML from '../../utils/stripHTML';
 import calcRotatedResizeOffset from '../../utils/calcRotatedResizeOffset';
 import {
   getBorderPositionCSS,
