+++ packages/url/README.md
@@ -1,0 +1,3 @@
+# Url
+
+URL utility package for Web Stories editor.
