+++ packages/url/src/url.js
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021 Google LLC
+ * Copyright 2022 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,6 +13,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
+export function toAbsoluteUrl(base, path) {
+  try {
+    return new URL(path, base).href;
+  } catch (error) {
+    return path;
+  }
+}
 
 export function isValidUrl(url) {
   try {
