+++ packages/wp-dashboard/package.json
@@ -37,6 +37,7 @@
     "@googleforcreators/react": "*",
     "@googleforcreators/story-editor": "*",
     "@googleforcreators/tracking": "*",
+    "@googleforcreators/url": "*",
     "@wordpress/api-fetch": "^6.0.1",
     "flagged": "^2.0.1",
     "opentype.js": "^1.3.4",
