+++ packages/wp-dashboard/src/api/authors.js
@@ -21,7 +21,7 @@ import apiFetch from '@wordpress/api-fetch';
 /**
  * External dependencies
  */
-import { addQueryArgs } from '@googleforcreators/design-system';
+import { addQueryArgs } from '@googleforcreators/url';
 
 export function getAuthors(config, search) {
   return apiFetch({
