+++ packages/wp-dashboard/src/api/publisherLogo.js
@@ -21,7 +21,7 @@ import apiFetch from '@wordpress/api-fetch';
 /**
  * External dependencies
  */
-import { addQueryArgs } from '@googleforcreators/design-system';
+import { addQueryArgs } from '@googleforcreators/url';
 
 /**
  * Fetch publisher logos.
