+++ packages/wp-story-editor/package.json
@@ -35,6 +35,7 @@
     "@googleforcreators/react": "*",
     "@googleforcreators/story-editor": "*",
     "@googleforcreators/tracking": "*",
+    "@googleforcreators/url": "*",
     "@wordpress/api-fetch": "^6.0.1",
     "flagged": "^2.0.1",
     "polished": "^4.1.4",
