+++ packages/wp-story-editor/src/api/hotlinkInfo.js
@@ -17,7 +17,7 @@
 /**
  * External dependencies
  */
-import { addQueryArgs } from '@googleforcreators/design-system';
+import { addQueryArgs } from '@googleforcreators/url';
 
 /**
  * WordPress dependencies
